Fixed categories woo route.
Added cron for:
Woo orders - checks every 3 hours if all orders from woo are in the app, if not -> get them manually.

TODO - Woo products - check every 3 hours if all products exist in woo, if not create them. Also checks if product price and qty is same in app and woo, if not - edit it in woo.